+++ Sources/RoktUXHelper/Data/Model/Event/RoktUXEvent.swift
@@ -10,6 +10,7 @@
 //  You may obtain a copy of the License at https://rokt.com/sdk-license-2-0/
 
 import Foundation
+import DcuiSchema
 
 public class RoktUXEvent {
 
@@ -181,7 +182,7 @@ public class RoktUXEvent {
         }
     }
 
-    public class CartItemStripePay: RoktUXEvent {
+    public class CartItemDevicePay: RoktUXEvent {
         public let layoutId: String
         public let name: String
         public let cartItemId: String
@@ -193,6 +194,7 @@ public class RoktUXEvent {
         public let quantity: Decimal
         public let totalPrice: Decimal?
         public let unitPrice: Decimal?
+        public let paymentProvider: PaymentProvider
 
         init(layoutId: String,
              name: String,
@@ -204,7 +206,8 @@ public class RoktUXEvent {
              providerData: String,
              quantity: Decimal,
              totalPrice: Decimal?,
-             unitPrice: Decimal?) {
+             unitPrice: Decimal?,
+             paymentProvider: PaymentProvider) {
             self.layoutId = layoutId
             self.name = name
             self.cartItemId = cartItemId
@@ -216,6 +219,7 @@ public class RoktUXEvent {
             self.quantity = quantity
             self.totalPrice = totalPrice
             self.unitPrice = unitPrice
+            self.paymentProvider = paymentProvider
         }
     }
 }