+++ Sources/RoktUXHelper/Data/Model/PartnerEventHandler/RoktEventListenerType.swift
@@ -26,5 +26,5 @@ enum RoktEventListenerType: String, CaseIterable {
     case FirstPositiveEngagement
     case OpenUrl
     case CartItemInstantPurchase
-    case CartItemStripePay
+    case CartItemDevicePay
 }